pyproject.toml
[tool.poetry]
name = "custom_sales_api"
version = "0.1.0"
description = ""
authors = ["bjgman12 <bjgman38@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
django-environ = "^0.4.5"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
